+++ src/main/kotlin/io/unthrottled/amii/listeners/IdleEventListener.kt
@@ -1,10 +1,11 @@
 package io.unthrottled.amii.listeners
 
-import com.intellij.ide.IdeEventQueue
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
+import com.intellij.util.Alarm
+import com.intellij.util.ui.UIUtil
 import io.unthrottled.amii.config.Config
 import io.unthrottled.amii.config.ConfigListener
 import io.unthrottled.amii.config.ConfigListener.Companion.CONFIG_TOPIC
@@ -12,36 +13,46 @@ import io.unthrottled.amii.events.EVENT_TOPIC
 import io.unthrottled.amii.events.UserEvent
 import io.unthrottled.amii.events.UserEventCategory
 import io.unthrottled.amii.events.UserEvents
+import io.unthrottled.amii.tools.BalloonTools
 import io.unthrottled.amii.tools.PluginMessageBundle
 import io.unthrottled.amii.tools.runSafely
+import java.awt.AWTEvent
+import java.awt.Toolkit
+import java.awt.event.AWTEventListener
+import java.awt.event.InputEvent
 import java.util.concurrent.TimeUnit
 
-class IdleEventListener(private val project: Project) : Runnable, Disposable {
+class IdleEventListener(private val project: Project) : Runnable, Disposable, AWTEventListener {
   private val messageBus = ApplicationManager.getApplication().messageBus.connect()
   private val log = Logger.getInstance(this::class.java)
+  private val rootPane = BalloonTools.getIDEFrame(project).component
+  private var idleTimeout =
+    TimeUnit.MILLISECONDS.convert(
+      getCurrentTimoutInMinutes(),
+      TimeUnit.MINUTES
+    ).toInt()
+  private val idleAlarm = Alarm()
 
   init {
     val self = this
     messageBus.subscribe(
       CONFIG_TOPIC,
       ConfigListener { newPluginState ->
-        IdeEventQueue.getInstance().removeIdleListener(self)
-        IdeEventQueue.getInstance().addIdleListener(
-          self,
-          TimeUnit.MILLISECONDS.convert(
-            newPluginState.idleTimeoutInMinutes,
-            TimeUnit.MINUTES
-          ).toInt()
-        )
+        idleAlarm.cancelAllRequests()
+        self.idleTimeout = TimeUnit.MILLISECONDS.convert(
+          newPluginState.idleTimeoutInMinutes,
+          TimeUnit.MINUTES
+        ).toInt()
+        idleAlarm.addRequest(self, self.idleTimeout)
       }
     )
-    IdeEventQueue.getInstance().addIdleListener(
+    Toolkit.getDefaultToolkit().addAWTEventListener(
       this,
-      TimeUnit.MILLISECONDS.convert(
-        getCurrentTimoutInMinutes(),
-        TimeUnit.MINUTES
-      ).toInt()
+      AWTEvent.MOUSE_EVENT_MASK or
+        AWTEvent.MOUSE_MOTION_EVENT_MASK or
+        AWTEvent.KEY_EVENT_MASK
     )
+    idleAlarm.addRequest(this, idleTimeout)
   }
 
   private fun getCurrentTimoutInMinutes(): Long =
@@ -49,14 +60,15 @@ class IdleEventListener(private val project: Project) : Runnable, Disposable {
 
   override fun dispose() {
     messageBus.dispose()
+    idleAlarm.dispose()
     runSafely({
-      IdeEventQueue.getInstance().removeIdleListener(this)
+      Toolkit.getDefaultToolkit().removeAWTEventListener(this)
     }) {}
   }
 
   override fun run() {
     log.debug("Observed idled timeout")
-    ApplicationManager.getApplication().messageBus
+    project.messageBus
       .syncPublisher(EVENT_TOPIC)
       .onDispatch(
         UserEvent(
@@ -66,5 +78,13 @@ class IdleEventListener(private val project: Project) : Runnable, Disposable {
           project
         )
       )
+    idleAlarm.addRequest(this, idleTimeout)
+  }
+
+  override fun eventDispatched(e: AWTEvent) {
+    if (e !is InputEvent || !UIUtil.isDescendingFrom(e.component, rootPane)) return
+
+    idleAlarm.cancelAllRequests()
+    idleAlarm.addRequest(this, idleTimeout)
   }
 }