+++ docs/RELEASE-NOTES.md
@@ -4,4 +4,5 @@
 
 
 ### Changed
-- How the AFK gifs are dismissed. They only get dismissed when the project it is on gains focus again. Not when any project gains focus.
+- The number of available MIKUs. Each project gets it's very own MIKU.
+  They function independently of each other and have their own moods.